requirements.txt
django-allauth==0.38.0
Django==2.1.4
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
pytz==2018.7
six==1.12.0
psycopg2==2.7.6
